separate css and js from rust
