+++ src/rustbook/main.rs
@@ -35,7 +35,6 @@ mod build;
 mod serve;
 mod test;
 
-mod css;
 mod javascript;
 
 static EXIT_STATUS: AtomicIsize = ATOMIC_ISIZE_INIT;
