+++ src/rustbook/static/rustbook.css
@@ -1,16 +1,3 @@
-// Copyright 2014 The Rust Project Developers. See the COPYRIGHT
-// file at the top-level directory of this distribution and at
-// http://rust-lang.org/COPYRIGHT.
-//
-// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
-// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
-// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
-// option. This file may not be copied, modified, or distributed
-// except according to those terms.
-
-// The rust-book CSS in string form.
-
-pub static STYLE: &'static str = r#"
 @import url("../rust.css");
 
 body {
@@ -57,7 +44,7 @@ h1, h2, h3, h4, h5, h6 {
 }
 
 @media only print {
-    #toc, #nav {
+    #toc, #nav, #menu-bar {
         display: none;
     }
 }
@@ -174,4 +161,3 @@ pre {
 .right {
     float: right;
 }
-"#;
